a few another improvements...
Can be connected to model 3 trucks from separate file. Defining files
and the relative positions of lift trucks in .mmd.
assigning to an object of T3dmodel of information about him - tris,
submodels count, inc file, model file
display of the above information, the closer to the object model.
assigning to object vehicle information called the station on which it
is located and the number of the track (taken from the properties of the
track)
possibility of any crossover crossover switch at a distance of 2m from
the camera (ctrl + z)